put helper functions in Helper file
