+++ package.json
@@ -16,7 +16,8 @@
     "react-d3-tree": "^3.5.2",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.9.0",
-    "react-scripts": "^5.0.1"
+    "react-scripts": "^5.0.1",
+    "uuid": "^9.0.0"
   },
   "scripts": {
     "predeploy": "npm run build",
@@ -24,7 +25,8 @@
     "start": "craco start",
     "build": "craco build",
     "test": "craco test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "lint": "eslint --ext .js --ignore-path .gitignore ."
   },
   "eslintConfig": {
     "extends": [
